requirements.txt
# Django and DRF
Django==5.1.5
django-configurations==2.5.1
djangorestframework==3.15.2
django-filter==24.3

# Gunicorn
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# AWS / Storages
boto3==1.35.76
django-storages==1.14.4

# Dev & Testing
pytest-django==4.9.0
factory-boy==3.3.1
pytest==7.4.2
coverage==7.6.9
ipdb==0.13.13
ipython==8.30.0
flake8==7.1.1
mkdocs==1.6.1

# Utility
setuptools==75.6.0

django-simple-history==3.3.0
django-celery-results
redis==5.0.0
requests==2.31.0

django-allauth==0.52.0
sentry-sdk==1.29.2
drf-spectacular==0.26.3

# Celery for async tasks
celery==5.3.4

black==23.7.0
isort==5.12.0